chore: Update package-lock.json and package.json with cors dependency
This commit adds the cors dependency to the package-lock.json and package.json files. The cors package is updated to version 2.8.5. This change is necessary to enable Cross-Origin Resource Sharing (CORS) in the application, allowing it to make requests from different domains.

Note: This commit message follows the established convention of using the "chore" type for updates to package dependencies.